rework of game systems so there isnt circular dependencies
