+++ src/game/GameObjects/GameGeometry/DebugLevel/Ceilings.h
@@ -3,10 +3,11 @@
 #include <stdlib.h>
 #include <stdio.h>
 
+#include "game/GameConstants.h"
+
 #include "math/Vector.h"
 #include "game/Game.h"
 #include "math/mathUtils.h"
-#include "game/GameConstants.h"
 
 OnStart OnCeiling1Start(GameObject* gameObject);
 OnStart OnCeiling2Start(GameObject* gameObject);
