+++ src/game/GameObjects/GameGeometry/DebugLevel/Floors.h
@@ -3,10 +3,10 @@
 #include <stdlib.h>
 #include <stdio.h>
 
+#include "game/GameConstants.h"
 #include "math/Vector.h"
 #include "game/Game.h"
 #include "math/mathUtils.h"
-#include "game/GameConstants.h"
 
 OnStart OnFloor1Start(GameObject* gameObject);
 OnStart OnFloor2Start(GameObject* gameObject);
