+++ src/gui/Gui.c
@@ -95,10 +95,10 @@ void RigidBodyWidget(RigidBody* rigidBody)
 {
 	if (igCollapsingHeader_TreeNodeFlags("RigidBody", ImGuiTreeNodeFlags_CollapsingHeader))
 	{
+		igCheckbox("Bounding Box View", &rigidBody->debug);
 		igInputFloat3("Bounding Box Min", &rigidBody->boundingBox.minPos, "%.02f", ImGuiInputTextFlags_None);
 		igInputFloat3("Bounding Box Max", &rigidBody->boundingBox.maxPos, "%.02f", ImGuiInputTextFlags_None);
 		igInputFloat3("Velocity", &rigidBody->velocity, "%.02f", ImGuiInputTextFlags_None);
-		igCheckbox("Is Colliding", &rigidBody->isColliding); 
 		igTreePop();
 		igSeparator();
 	}
